requirements.txt
# 测试核心
pytest>=9.0
allure-pytest>=2.15
allure-python-commons>=2.15
allure-combine>=1.0

# 接口/UI自动化
requests>=2.33
playwright>=1.58

# 数据处理
pyyaml>=6.0